@@ -1,7 +1,7 @@
 package com.andr.common.tool.log;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.util.Log;
 
 import com.orhanobut.logger.AndroidLogAdapter;
