@@ -2,7 +2,7 @@ package com.andr.view.dialog;
 
 import android.app.Dialog;
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.view.KeyEvent;
 import android.view.WindowManager;
 import android.widget.TextView;
