@@ -1,8 +1,8 @@
 package com.andr.common.tool.net.okhttpUtil.builder;
 
 
-import com.andr.common.tool.net.okhttpUtil.request.HttpRequestCall;
 import com.andr.common.tool.net.okhttpUtil.request.DownFileRequest;
+import com.andr.common.tool.net.okhttpUtil.request.HttpRequestCall;
 
 /**
  * <pre>
